Fix comment problems

+++ workflows/services/docker.py
@@ -11,6 +11,13 @@ from workflows.models import DockerTaskExecution, TaskExecution
 from workflows.services.task import TaskService
 from django.conf import settings
 
+def add_file_to_datasets(file_path, filename, task_execution):
+    user = task_execution.task.creator
+    task_name = task_execution.task.name
+    path_in_media = "datasets/" + user.username + "/"
+    os.makedirs(settings.MEDIA_ROOT + path_in_media, exist_ok=True)
+    os.replace(file_path, settings.MEDIA_ROOT + path_in_media + filename)
+    Dataset.objects.create(creator=user, file=path_in_media + filename, title=f"created_by_{task_name}")
 
 class DockerTaskService(TaskService):
     accessible_datasets_path = '/datasets/'
@@ -45,43 +52,39 @@ class DockerTaskService(TaskService):
                 status = TaskExecution.StatusChoices.FAILED
             else:
                 status = TaskExecution.StatusChoices.SUCCESS
-                self._exctract_data_from_execution(task_execution)
+                self.exctract_data_from_execution(task_execution)
 
             task_execution.status = status
             task_execution.save()
 
         return status.name
 
-    def _add_file_to_datasets(self, file_path, filename, task_execution):
-        user = task_execution.task.creator
-        task_name = task_execution.task.name
-        path_in_media = "datasets/" + user.username + "/"
-        os.makedirs(settings.MEDIA_ROOT + path_in_media, exist_ok=True)
-        os.replace(file_path, settings.MEDIA_ROOT + path_in_media + filename)
-        Dataset.objects.create(creator=user, file=path_in_media + filename, title=f"created_by_{task_name}")
-
-    def _exctract_data_from_execution(self, task_execution):
+    def _recieve_data_from_container(self, task_execution, extract_path):
         client = docker.from_env()
         container = client.containers.get(task_execution.dockertaskexecution.container_id)
         f = BytesIO()
         try:
             bits, stat = container.get_archive(self.container_extract_datasets_path)
         except:
-            # There is no data to be extracted
-            return
+            return False
         for chunk in bits:
             f.write(chunk)
         f.seek(0)
         tar = tarfile.open(fileobj=f)
-        task_execution_id = str(task_execution.id)
-        extract_path = f"./task_{task_execution_id}"
         tar.extractall(path=extract_path)
         tar.close()
         f.close()
+        return True
+
+    def exctract_data_from_execution(self, task_execution):
+        extract_path = f"./task_{task_execution.id}"
+        if (not self._recieve_data_from_container(task_execution, extract_path)):
+            # There is no data to be extracted
+            return
         extracted_files_dir = extract_path + self.container_extract_datasets_path
         for file in os.listdir(extracted_files_dir):
             if (not is_file_valid(file)):
                 print(f'{file} is not a supported type')
             file_path = os.path.abspath(extracted_files_dir + file)
-            self._add_file_to_datasets(file_path, file, task_execution)
+            add_file_to_datasets(file_path, file, task_execution)
         shutil.rmtree(extract_path, ignore_errors=True)
